requirements.txt
streamlit
crewai
openai
langchain
tavily-python
python-dotenv
chromadb>=0.4.22